Thing: Separate tagging directions.
Things maintained a single tag list, representing edges between things.
However, tagging a thing with something has a different meaning than
a thing tagging some other thing. For example, there's a difference
between tagging 'banana' with 'food'... and tagging 'food' with 'banana'.
Furthermore, we want to differentiate and show, in a shell, the things
that 'Thing A' is tagged with vs the things that 'Thing A' tags.

Add the concept of 'targets' to Thing, which represent the things that
a Thing tags. Regardless of the tagging direction, add the appropriate
URIs to both the tags and targets lists.

At this point, a Thing maintains two neighbor lists: one for the things
that tag it, and one for the things that it tags.